Increase GHC compat range
This makes a new version of the package, 0.1.0.1, that has compatibility
with a larger range of GHC versions - in particular, it support GHC 9.2
with its breaking changes to template-haskell in the `ConP` constructor.

It also fixes a deprecation warning w.r.t the use of `*` as a type -
this should be compatible with all supported versions of GHC, since
`Type` was introduced in `base-4.9` with GHC 8.

The new GHC versions required a Nix bump, which also means that
`stdenv.lib` is no longer supported.